fix CalcPi precision

+++ CalcPi.java
@@ -4,7 +4,7 @@ public class CalcPi {
 	int N = Integer.parseInt(args[0]);
 	double sum = 0.0;
 	for (int i = 0; i < N; i++) {
-		double p = (1.0/((2 * i) + 1));
+		double p = (1.0/((double)(2 * i) + 1));
 		
 		if(i % 2 == 0) { 
 			sum = sum + p;
@@ -16,6 +16,6 @@ public class CalcPi {
 		double pi = 4 * sum;
 		System.out.println(pi);
 
-		}
+		 }
 	}
